requirements.txt
faiss-cpu
streamlit
sentence-transformers
numpy
ollama
pandas
beautifulsoup4
requests
PyPDF2
lxml
tqdm